[COMMON] media: mfc: fix prevent defect
This patch fixes about prevent defect.
- Dereference after null check
- Uninitialized pointer read
- Unsigned compared against 0
- Unused value
- Logically dead code
- Operands don't affect result
- Unchecked return value

+++ drivers/media/platform/exynos/mfc/s5p_mfc_buf.c
@@ -1018,7 +1018,7 @@ int s5p_mfc_set_enc_ref_buffer(struct s5p_mfc_ctx *ctx)
 	struct s5p_mfc_dev *dev = ctx->dev;
 	struct s5p_mfc_enc *enc = ctx->enc_priv;
 	dma_addr_t buf_addr1;
-	size_t buf_size1;
+	int buf_size1;
 	int i;
 
 	mfc_debug_enter();
@@ -1026,7 +1026,7 @@ int s5p_mfc_set_enc_ref_buffer(struct s5p_mfc_ctx *ctx)
 	buf_addr1 = ctx->codec_buf_phys;
 	buf_size1 = ctx->codec_buf_size;
 
-	mfc_debug(2, "Buf1: %p (%ld)\n", (void *)buf_addr1, buf_size1);
+	mfc_debug(2, "Buf1: %p (%d)\n", (void *)buf_addr1, buf_size1);
 
 	/* start address of per buffer is aligned */
 	for (i = 0; i < ctx->dpb_count; i++) {
@@ -1058,7 +1058,7 @@ int s5p_mfc_set_enc_ref_buffer(struct s5p_mfc_ctx *ctx)
 		buf_size1 -= enc->tmv_buffer_size;
 	}
 
-	mfc_debug(2, "Buf1: %p, buf_size1: %ld (ref frames %d)\n",
+	mfc_debug(2, "Buf1: %p, buf_size1: %d (ref frames %d)\n",
 			(void *)buf_addr1, buf_size1, ctx->dpb_count);
 	if (buf_size1 < 0) {
 		mfc_debug(2, "Not enough memory has been allocated.\n");
@@ -1173,7 +1173,7 @@ int s5p_mfc_set_dec_frame_buffer(struct s5p_mfc_ctx *ctx)
 	struct s5p_mfc_dec *dec;
 	unsigned int i, frame_size_mv;
 	dma_addr_t buf_addr1;
-	size_t buf_size1;
+	int buf_size1;
 	int align_gap;
 	struct s5p_mfc_buf *buf;
 	struct s5p_mfc_raw_info *raw, *tiled_ref;
@@ -1205,7 +1205,7 @@ int s5p_mfc_set_dec_frame_buffer(struct s5p_mfc_ctx *ctx)
 	buf_addr1 = ctx->codec_buf_phys;
 	buf_size1 = ctx->codec_buf_size;
 
-	mfc_debug(2, "Buf1: %p (%ld)\n", (void *)buf_addr1, buf_size1);
+	mfc_debug(2, "Buf1: %p (%d)\n", (void *)buf_addr1, buf_size1);
 	mfc_debug(2, "Total DPB COUNT: %d\n", dec->total_dpb_count);
 	mfc_debug(2, "Setting display delay to %d\n", dec->display_delay);
 
@@ -1316,14 +1316,14 @@ int s5p_mfc_set_dec_frame_buffer(struct s5p_mfc_ctx *ctx)
 			align_gap = buf_addr1 - align_gap;
 			buf_size1 -= align_gap;
 
-			mfc_debug(2, "\tBuf1: %p, size: %ld\n", (void *)buf_addr1, buf_size1);
+			mfc_debug(2, "\tBuf1: %p, size: %d\n", (void *)buf_addr1, buf_size1);
 			MFC_WRITEL(buf_addr1, S5P_FIMV_D_MV_BUFFER0 + i * 4);
 			buf_addr1 += frame_size_mv;
 			buf_size1 -= frame_size_mv;
 		}
 	}
 
-	mfc_debug(2, "Buf1: %p, buf_size1: %ld (frames %d)\n",
+	mfc_debug(2, "Buf1: %p, buf_size1: %d (frames %d)\n",
 			(void *)buf_addr1, buf_size1, dec->total_dpb_count);
 	if (buf_size1 < 0) {
 		mfc_debug(2, "Not enough memory has been allocated.\n");
